install the script as "invokeai-batch"
